Minimal changes to enable alternate endpoints
The FISS API operations were always performed on the default
firecloud API endpoint ("https://api.firecloud.org/api/") even when
an alternate firecloud API was configured.

This was happening because the FISS API functions to perform
HTTP operations had signatures like this:
```
def __get(methcall, headers=None, root_url=fcconfig.root_url, **kwargs):
```
and these are generally called without passing a value for `root_url`.
The problem with this is that Python default parameter values are only
evaluated at the time the function is defined, which for the FISS API is
at the time the API module is imported.
Thus, the `root_url` value defined at import time, which is the default
value ("https://api.firecloud.org/api/") was *always* used.

This commit contains a minimal set of changes that enable the
FISS API to work  with alternate endpoints. There may (or may not)
be other, better ways of making this change, and the change in this
commit has not been verified to cover all FISS functionality.